# Avoid single character `String` allocations when appending characters