Removed calls to `String::chr()` when appending characters to `Strings` in `Expression`, `Resource`, and `VariantParser`, to avoid creating temporary `Strings` for each character.  Also updated the `Resource` case to resize `String` up front, since size is known.

This makes `Expression.parse` around 50% faster, `Resource.generate_scene_unique_id` around 5x faster, and `ConfigFile.load` around 25% faster, compared with gdscript below:

```gdscript
func _ready() -> void:
	run_test("expression_parse")
	run_test("resource_unique_id")
	run_test("load_config")
	
func run_test(p_name : String):
	var start := Time.get_ticks_msec()
	call(p_name)
	var end := Time.get_ticks_msec()
	print("%s: %dms" % [p_name, end - start])

func expression_parse():
	var expression := Expression.new()
	for i in 100000:
		expression.parse("12345 + 67890 - sqrt(1000000)")

func resource_unique_id():
	for i in 1000000:
		Resource.generate_scene_unique_id()

func load_config():
	for i in 1000:
		var config := ConfigFile.new()
		config.load("res://icon.svg.import")
```

old:
```
expression_parse: 380ms
resource_unique_id: 447ms
load_config: 118ms
```

new:
```
expression_parse: 253ms
resource_unique_id: 80ms
load_config: 91ms
```

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->